Isolate Mavlink Implementations into new Library
