+++ src/AnalyzeView/MAVLinkConsoleController.h
