+++ src/comm/Mavlink/MAVLinkProtocol.h
@@ -1,15 +1,5 @@
-/****************************************************************************
- *
- * (c) 2009-2020 QGROUNDCONTROL PROJECT <http://www.qgroundcontrol.org>
- *
- * QGroundControl is licensed according to the terms in the file
- * COPYING.md in the root of the source code directory.
- *
- ****************************************************************************/
-
 #pragma once
 
-#include "LinkInterface.h"
 #include "QGCMAVLink.h"
 #include "QGCTemporaryFile.h"
 #include "QGCToolbox.h"
@@ -18,6 +8,7 @@
 #include <QtCore/QByteArray>
 #include <QtCore/QLoggingCategory>
 
+class LinkInterface;
 class LinkManager;
 class MultiVehicleManager;
 class QGCApplication;
